% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecomon.R
\name{as_ecomon}
\alias{as_ecomon}
\title{Cast staged data for one species into an ecomon compatible data frame.}
\usage{
as_ecomon(
  x = read_calanus(),
  species = c("calanus_finmarchicus", "calanus_glacialis", "calanus_hyperboreus")[1]
)
}
\arguments{
\item{x}{data frame of AZMP data}

\item{species}{chr, one of "calanus_finmarchicus", "calanus_glacialis",
or "calanus_hyperboreus"}
}
\value{
data frame, suitable to bind to ecomon (corrected) staged data.
See \code{\link[ecomon:read_staged]{ecomon::read_staged()}} for more information on the output.
}
\description{
Note that not everything transfers 1:1, so we make some choices for some
of the ecomon values (\emph{e.g.} NA for gear_volume_filtered)
}
