% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecomon.R
\name{ecomon_template}
\alias{ecomon_template}
\title{Retrieve a template to match the ecomon staged data table}
\usage{
ecomon_template()
}
\value{
zero row data frame
}
\description{
Retrieve a template to match the ecomon staged data table
}
